Fix: Move data tag logic away from the component to prevent multiple renders [APP-1502]
